refactor: rename response types to just responses
