+++ src/lib/api_app/src/api/routes/mod.rs
@@ -1,5 +1,6 @@
 use super::{
-    response_types::{APIRoutingResponse, APIRoutingError, ParsedRequest},
+    responses::{APIRoutingResponse, APIRoutingError},
+    utils::ParsedRequest,
 };
 
 pub mod application;
