+++ src/lib/api_app/src/api/routes/testbed/productizers/mod.rs
@@ -1,6 +1,9 @@
 use http::{HeaderMap, HeaderValue};
 
-use crate::api::response_types::{APIRoutingError, ParsedRequest};
+use crate::api::{
+    responses::APIRoutingError,
+    utils::ParsedRequest,
+};
 
 pub mod figure;
 pub mod job;
